requirements.txt

livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-silero


python-dotenv


numpy

edge-tts
groq
